+++ common-api/src/main/scala/pl/touk/nussknacker/engine/api/MetaData.scala
@@ -8,7 +8,7 @@ import pl.touk.nussknacker.engine.api.CirceUtil._
 
 @JsonCodec case class LayoutData(x: Long, y: Long)
 
-// todo: MetaData should hold ProcessName as id
+// TODO: MetaData should hold ProcessName as id
 @ConfiguredJsonCodec(encodeOnly = true) case class MetaData(id: String,
                                                             additionalFields: ProcessAdditionalFields) {
   def isFragment: Boolean = typeSpecificData.isFragment