+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/AppResourcesSpec.scala
@@ -2,57 +2,54 @@ package pl.touk.nussknacker.ui.api
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import com.typesafe.config.ConfigFactory
+import com.typesafe.config.{Config, ConfigFactory}
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
 import io.circe.Json
 import io.circe.syntax.EncoderOps
+import org.scalatest.exceptions.TestFailedException
 import org.scalatest.funsuite.AnyFunSuite
-import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, OptionValues}
 import org.scalatest.matchers.should.Matchers
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, OptionValues}
+import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus
 import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleProcessStateDefinitionManager, SimpleStateStatus}
 import pl.touk.nussknacker.engine.api.deployment.{ProcessState, StateStatus, StatusDetails}
 import pl.touk.nussknacker.engine.api.process.{EmptyProcessConfigCreator, ProcessName, VersionId}
 import pl.touk.nussknacker.engine.testing.LocalModelData
+import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 import pl.touk.nussknacker.engine.version.BuildInfo
 import pl.touk.nussknacker.test.PatientScalaFutures
 import pl.touk.nussknacker.ui.api.helpers.TestFactory.{emptyProcessingTypeDataProvider, mapProcessingTypeDataProvider, withAdminPermissions, withPermissions, withoutPermissions}
-import pl.touk.nussknacker.ui.api.helpers.{EspItTest, StubDeploymentService, TestFactory}
-import pl.touk.nussknacker.ui.process.processingtypedata.ProcessingTypeDataReload
-
-import scala.jdk.CollectionConverters._
-import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
+import pl.touk.nussknacker.ui.api.helpers.{NuItTest, NuScenarioConfigurationHelper, StubDeploymentService, TestFactory}
+import pl.touk.nussknacker.ui.process.processingtypedata.{ProcessingTypeDataProvider, ProcessingTypeDataReload}
 
 import java.util.Collections
+import scala.jdk.CollectionConverters._
 
-class AppResourcesSpec extends AnyFunSuite with ScalatestRouteTest with Matchers with PatientScalaFutures with FailFastCirceSupport
-  with OptionValues with BeforeAndAfterEach with BeforeAndAfterAll with EspItTest {
+class AppResourcesSpec
+  extends AnyFunSuite
+    with ScalatestRouteTest
+    with Matchers
+    with PatientScalaFutures
+    with FailFastCirceSupport
+    with OptionValues
+    with BeforeAndAfterEach
+    with BeforeAndAfterAll
+    with NuItTest // fixme: it's not fully used here yet
+    with NuScenarioConfigurationHelper {
 
-  private val emptyReload = new ProcessingTypeDataReload {
-    override def reloadAll(): Unit = ???
-  }
-
-  private def processStatus(status: StateStatus): ProcessState =
-    SimpleProcessStateDefinitionManager.processState(StatusDetails(status, None))
-
-  private def prepareBasicAppResources(statuses: Map[ProcessName, StateStatus], withConfigExposed: Boolean = true) = {
-    val deploymentService = new StubDeploymentService(statuses.mapValuesNow(processStatus))
-    new AppResources(ConfigFactory.empty(), emptyReload, emptyProcessingTypeDataProvider, futureFetchingProcessRepository,
-      TestFactory.processValidation, deploymentService, exposeConfig = withConfigExposed, processCategoryService
-    )
-  }
-
-  test("it should return healthcheck also if cannot retrieve statuses") {
+  test("it should return health check also if cannot retrieve statuses") {
     createDeployedProcess(ProcessName("id1"))
     createDeployedProcess(ProcessName("id2"))
     createDeployedProcess(ProcessName("id3"))
-    val statuses = Map(
-      ProcessName("id1") -> ProblemStateStatus.FailedToGet,
-      ProcessName("id2") -> SimpleStateStatus.Running,
-      ProcessName("id3") -> ProblemStateStatus.shouldBeRunning(VersionId(1L), "user"),
+    val resources = prepareBasicAppResources(
+      statuses = Map(
+        ProcessName("id1") -> ProblemStateStatus.FailedToGet,
+        ProcessName("id2") -> SimpleStateStatus.Running,
+        ProcessName("id3") -> ProblemStateStatus.shouldBeRunning(VersionId(1L), "user"),
+      )
     )
 
-    val resources = prepareBasicAppResources(statuses)
     val result = Get("/app/healthCheck/process/deployment") ~> withPermissions(resources, testPermissionRead)
 
     result ~> check {
@@ -65,11 +62,12 @@ class AppResourcesSpec extends AnyFunSuite with ScalatestRouteTest with Matchers
   test("it shouldn't return healthcheck when scenario canceled") {
     createDeployedCanceledProcess(ProcessName("id1"))
     createDeployedProcess(ProcessName("id2"))
-    val statuses = Map(
-      ProcessName("id2") -> ProblemStateStatus.shouldBeRunning(VersionId(1L), "user"),
+    val resources = prepareBasicAppResources(
+      statuses = Map(
+        ProcessName("id2") -> ProblemStateStatus.shouldBeRunning(VersionId(1L), "user"),
+      )
     )
 
-    val resources = prepareBasicAppResources(statuses)
     val result = Get("/app/healthCheck/process/deployment") ~> withPermissions(resources, testPermissionRead)
 
     result ~> check {
@@ -82,12 +80,13 @@ class AppResourcesSpec extends AnyFunSuite with ScalatestRouteTest with Matchers
   test("it should return healthcheck ok if statuses are ok") {
     createDeployedProcess(ProcessName("id1"))
     createDeployedProcess(ProcessName("id2"))
-    val statuses = Map(
-      ProcessName("id1") -> SimpleStateStatus.Running,
-      ProcessName("id2") -> SimpleStateStatus.Running,
+    val resources = prepareBasicAppResources(
+      statuses = Map(
+        ProcessName("id1") -> SimpleStateStatus.Running,
+        ProcessName("id2") -> SimpleStateStatus.Running,
+      )
     )
 
-    val resources = prepareBasicAppResources(statuses)
     val result = Get("/app/healthCheck/process/deployment") ~> withPermissions(resources, testPermissionRead)
 
     result ~> check {
@@ -97,11 +96,12 @@ class AppResourcesSpec extends AnyFunSuite with ScalatestRouteTest with Matchers
 
   test("it should not report deployment in progress as fail") {
     createDeployedProcess(ProcessName("id1"))
-    val statuses = Map(
-      ProcessName("id1") -> SimpleStateStatus.Running,
+    val resources = prepareBasicAppResources(
+      statuses = Map(
+        ProcessName("id1") -> SimpleStateStatus.Running,
+      )
     )
 
-    val resources = prepareBasicAppResources(statuses)
     val result = Get("/app/healthCheck/process/deployment") ~> withPermissions(resources, testPermissionRead)
 
     result ~> check {
@@ -110,7 +110,7 @@ class AppResourcesSpec extends AnyFunSuite with ScalatestRouteTest with Matchers
   }
 
   test("it should return config") {
-    val resources = prepareBasicAppResources(Map.empty)
+    val resources = prepareBasicAppResources(statuses = Map.empty)
 
     val result = Get("/app/config") ~> withAdminPermissions(resources)
 
@@ -120,7 +120,10 @@ class AppResourcesSpec extends AnyFunSuite with ScalatestRouteTest with Matchers
   }
 
   test("it shouldn't return config when disabled") {
-    val resources = prepareBasicAppResources(Map.empty, withConfigExposed = false)
+    val resources = prepareBasicAppResources(
+      statuses = Map.empty,
+      withConfigExposed = false
+    )
 
     val result = Get("/app/config") ~> withAdminPermissions(resources)
 
@@ -133,28 +136,63 @@ class AppResourcesSpec extends AnyFunSuite with ScalatestRouteTest with Matchers
     val creatorWithBuildInfo = new EmptyProcessConfigCreator {
       override def buildInfo(): Map[String, String] = Map("fromModel" -> "value1")
     }
-    val modelData = LocalModelData(ConfigFactory.empty(), creatorWithBuildInfo)
-
+    val modelData = LocalModelData(
+      inputConfig = ConfigFactory.empty(),
+      configCreator = creatorWithBuildInfo
+    )
     val globalConfig = Map("testConfig" -> "testValue", "otherConfig" -> "otherValue")
+    val resources = prepareBasicAppResources(
+      config = ConfigFactory.parseMap(Collections.singletonMap("globalBuildInfo", globalConfig.asJava)),
+      statuses = Map.empty,
+      modelData = mapProcessingTypeDataProvider("test1" -> modelData),
+      withConfigExposed = false
+    )
 
-    val resources = new AppResources(ConfigFactory.parseMap(Collections.singletonMap("globalBuildInfo", globalConfig.asJava)), emptyReload,
-      mapProcessingTypeDataProvider("test1" -> modelData), futureFetchingProcessRepository, TestFactory.processValidation,
-      deploymentService, exposeConfig = false, processCategoryService)
+    val expectedEntity = (BuildInfo.toMap.mapValuesNow(_.toString) ++ globalConfig).mapValuesNow(_.asJson) +
+      ("processingType" -> Map("test1" -> creatorWithBuildInfo.buildInfo()).asJson)
 
-    val result = Get("/app/buildInfo") ~> withoutPermissions(resources)
-    result ~> check {
+    Get("/app/buildInfo") ~> withoutPermissions(resources) ~> check {
       status shouldBe StatusCodes.OK
-      entityAs[Map[String, Json]] shouldBe (BuildInfo.toMap.mapValuesNow(_.toString) ++ globalConfig).mapValuesNow(_.asJson) + ("processingType" -> Map("test1" -> creatorWithBuildInfo.buildInfo()).asJson)
+      entityAs[Map[String, Json]] shouldBe (expectedEntity)
     }
   }
 
   test("it should should return simple designer health check (not checking scenario statuses) without authentication") {
-    val resources = prepareBasicAppResources(Map.empty)
+    val resources = prepareBasicAppResources(statuses = Map.empty)
 
     val result = Get("/app/healthCheck") ~> withoutPermissions(resources)
+
     result ~> check {
       status shouldBe StatusCodes.OK
     }
+  }
+
+  private val emptyReload = new ProcessingTypeDataReload {
+    override def reloadAll(): Unit = throw new TestFailedException(
+      message = Some("Should not be called in this test case"),
+      cause = None,
+      failedCodeStackDepth = 0
+    )
+  }
+
+  private def processStatus(status: StateStatus): ProcessState =
+    SimpleProcessStateDefinitionManager.processState(StatusDetails(status, None))
+
+  private def prepareBasicAppResources(config: Config = ConfigFactory.empty(),
+                                       statuses: Map[ProcessName, StateStatus],
+                                       modelData: ProcessingTypeDataProvider[ModelData, _] = emptyProcessingTypeDataProvider,
+                                       withConfigExposed: Boolean = true) = {
+    val deploymentService = new StubDeploymentService(statuses.mapValuesNow(processStatus))
+    new AppResources(
+      config = config,
+      processingTypeDataReload = emptyReload,
+      modelData = modelData,
+      processRepository = futureFetchingProcessRepository,
+      processValidation = TestFactory.processValidation,
+      deploymentService = deploymentService,
+      exposeConfig = withConfigExposed,
+      processCategoryService = processCategoryService
+    )
   }
 
 }